requirements.txt
fastapi

uvicorn

psycopg2-binary

python-dotenv

pydantic

langchain
langchain-google-genai
streamlit

requests

pandas